AI system/requirements.txt
streamlit
transformers
torch
pdfminer.six

